feat: moved ethers_helpers tests to its own test file
